+++ prefect_airbyte/configuration.py
@@ -1,15 +1,18 @@
 """Tasks for updating and fetching Airbyte configurations"""
-from prefect import task
-from prefect.logging.loggers import get_logger
+from typing import Optional
+from warnings import warn
 
-from prefect_airbyte.client import AirbyteClient
+from prefect import get_run_logger, task
+
+from prefect_airbyte.server import AirbyteServer
 
 
 @task
 async def export_configuration(
-    airbyte_server_host: str = "localhost",
-    airbyte_server_port: int = "8000",
-    airbyte_api_version: str = "v1",
+    airbyte_server: Optional[AirbyteServer] = None,
+    airbyte_server_host: Optional[str] = None,
+    airbyte_server_port: Optional[int] = None,
+    airbyte_api_version: Optional[str] = None,
     timeout: int = 5,
 ) -> bytes:
 
@@ -17,10 +20,16 @@ async def export_configuration(
     Prefect Task that exports an Airbyte configuration via
     `{airbyte_server_host}/api/v1/deployment/export`.
 
+    As of `prefect-airbyte==0.1.3`, the kwargs `airbyte_server_host` and
+    `airbyte_server_port` can be replaced by passing an `airbyte_server` block
+    instance to generate the `AirbyteClient`. Using the `airbyte_server` block is
+    preferred, but the individual kwargs remain for backwards compatibility.
+
     Args:
-        airbyte_server_host: Airbyte instance hostname where connection is configured.
-        airbyte_server_port: Port where Airbyte instance is listening.
-        airbyte_api_version: Version of Airbyte API to use to export configuration.
+        airbyte_server: An `AirbyteServer` block for generating an `AirbyteClient`.
+        airbyte_server_host: Airbyte server host to connect to.
+        airbyte_server_port: Airbyte server port to connect to.
+        airbyte_api_version: Airbyte API version to use.
         timeout: Timeout in seconds on the `httpx.AsyncClient`.
 
     Returns:
@@ -35,6 +44,7 @@ async def export_configuration(
 
         from prefect import flow, task
         from prefect_airbyte.configuration import export_configuration
+        from prefect_airbyte.server import AirbyteServer
 
         @task
         def zip_and_write_somewhere(
@@ -50,9 +60,7 @@ async def export_configuration(
             # Run other tasks and subflows here
 
             airbyte_config = export_configuration(
-                    airbyte_server_host="localhost",
-                    airbyte_server_port="8000",
-                    airbyte_api_version="v1",
+                airbyte_server=AirbyteServer.load("oss-airbyte")
             )
 
             zip_and_write_somewhere(airbyte_config=airbyte_config)
@@ -60,17 +68,38 @@ async def export_configuration(
         example_trigger_sync_flow()
         ```
     """
+    logger = get_run_logger()
 
-    logger = get_logger()
+    if not airbyte_server:
+        warn(
+            "The use of `airbyte_server_host`, `airbyte_server_port`, and "
+            "`airbyte_api_version` is deprecated and will be removed in a "
+            "future release. Please pass an `airbyte_server` block to this "
+            "task instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        if any([airbyte_server_host, airbyte_server_port, airbyte_api_version]):
+            airbyte_server = AirbyteServer(
+                server_host=airbyte_server_host or "localhost",
+                server_port=airbyte_server_port or 8000,
+                api_version=airbyte_api_version or "v1",
+            )
+        else:
+            airbyte_server = AirbyteServer()
+    else:
+        if any([airbyte_server_host, airbyte_server_port, airbyte_api_version]):
+            logger.info(
+                "Ignoring `airbyte_server_host`, `airbyte_api_version`, "
+                "and `airbyte_server_port` because `airbyte_server` block "
+                " was passed. Using API URL from `airbyte_server` block: "
+                f"{airbyte_server.base_url!r}."
+            )
 
-    airbyte_base_url = (
-        f"http://{airbyte_server_host}:"
-        f"{airbyte_server_port}/api/{airbyte_api_version}"
-    )
+    async with airbyte_server.get_client(
+        logger=logger, timeout=timeout
+    ) as airbyte_client:
 
-    airbyte = AirbyteClient(logger, airbyte_base_url, timeout=timeout)
+        logger.info("Initiating export of Airbyte configuration")
 
-    logger.info("Initiating export of Airbyte configuration")
-    airbyte_config = await airbyte.export_configuration()
-
-    return airbyte_config
+        return await airbyte_client.export_configuration()